+++ cea/plots/demand/energy_final_use.py
@@ -13,7 +13,7 @@ from cea.plots.variable_naming import NAMING, LOGO, COLOR
 
 class EnergySupplyPlot(cea.plots.demand.energy_demand.EnergyDemandDistrictPlot):
     """Implement the energy-supply plot, inherits most of it's functionality from EnergyDemandDistrictPlot"""
-    name = "Energy Supply"
+    name = "Energy Use"
 
     def __init__(self, project, parameters, cache):
         super(EnergySupplyPlot, self).__init__(project, parameters, cache)
@@ -25,7 +25,7 @@ class EnergySupplyPlot(cea.plots.demand.energy_demand.EnergyDemandDistrictPlot):
     @property
     def layout(self):
         return go.Layout(barmode='stack',
-                         yaxis=dict(title='Energy Supply [MWh/yr]'),
+                         yaxis=dict(title='Energy Demand [MWh/yr]'),
                          xaxis=dict(title='Building Name'), showlegend=True)
